requirements.txt
flask
SQLAlchemy
flask-sqlalchemy
flask-login
flask-wtf
werkzeug
python-dotenv
email-validator
pytz